fix(cache): #4438 디스크 캐시 카운터를 usize atomic 으로 — 32-bit 타겟 빌드 회귀
graph.disk_load_hits / disk_cache.put_seq 가 Atomic(u64) 라 32-bit 타겟(wasm32-wasi,
x86-windows-msvc)에서 @atomicRmw 가 64-bit 정수를 거부 → atomic.zig:53 컴파일 실패
(CI 의 WASM Build / Publish Smoke / NAPI win32-ia32 job 동시 실패).

워드 크기 usize atomic 으로 전환: 32-bit=u32(atomic 지원), 64-bit=u64(기존과 동일),
단조 카운터/hit 카운터 의미 유지. minify.zig dead_toplevel_count 와 동일 선례.

+++ src/bundler/disk_cache.zig
@@ -32,7 +32,8 @@ const std = @import("std");
 
 /// put 마다 1씩 증가하는 프로세스 전역 단조 카운터. 같은 프로세스 내 모든 put 의 tmp 이름을
 /// 유일하게 만든다(atomic fetchAdd 라 두 put 이 같은 값을 받지 않음).
-var put_seq = std.atomic.Value(u64).init(0);
+/// usize: 32-bit 타겟(wasm32/x86-win)은 64-bit atomic 을 지원하지 않으므로 워드 크기 카운터 사용.
+var put_seq = std.atomic.Value(usize).init(0);
 
 pub const DiskCache = struct {
     allocator: std.mem.Allocator,
